moving count down to component folder
